requirements.txt
torch>=2.1.0
transformers>=4.57.0
accelerate>=0.31.0
tqdm>=4.66.0
shortuuid>=1.0.11
pillow>=10.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
rouge>=1.0.1
# Optional: bitsandbytes for 4-bit/8-bit loading
bitsandbytes>=0.43.0
